+++ VFS/source/NetDropbox/Host.h
@@ -30,7 +30,7 @@ public:
 
     virtual int Stat(const char *_path,
                      VFSStat &_st,
-                     int _flags,
+                     unsigned long _flags,
                      const VFSCancelChecker &_cancel_checker) override;
     
     virtual int Unlink(const char *_path,
@@ -45,7 +45,7 @@ public:
 
     virtual int FetchDirectoryListing(const char *_path,
                                       shared_ptr<VFSListing> &_target,
-                                      int _flags,
+                                      unsigned long _flags,
                                       const VFSCancelChecker &_cancel_checker) override;
 
     virtual int CreateFile(const char* _path,
